+++ app/utils/defaultSettings.js
@@ -27,6 +27,7 @@ module.exports = {
   mainColor: '#478484',
   miniBreakColor: '#478484',
   transparentMode: false,
+  blurredBackground: false,
   opacity: 0.9,
   audio: 'crystal-glass',
   miniBreakAudio: 'crystal-glass',